Refactor `Sidebar.MenuItem` style
