+++ src/common/components/generics/Sidebar/Sidebar.module.scss
@@ -47,25 +47,19 @@
     position: relative;
     display: flex;
     flex-direction: column;
-    gap: 30px;
-    padding-left: 1rem;
+    gap: 40px;
+    padding-left: 20px;
     list-style: none;
 
     &__item {
-      &__link {
-        position: relative;
-        display: flex;
-        align-items: center;
-        gap: 10px;
-        padding: 10px 20px;
+      position: relative;
 
-        &:hover {
-          @include sidebarItem;
-        }
+      &:hover {
+        @include sidebarItem;
+      }
 
-        &_active {
-          @include sidebarItem;
-        }
+      &_active {
+        @include sidebarItem;
       }
     }
